Improvements to Auth.net Refunds (DROP-2424) (#1517)
* Add messages, void transaction

* Fix refund calculation

* Fix Order Status

* Fix tests

* Fix failing test

* Add status to Refunds

* Minor convention change

* Minor fixes.

* Fix flake8.

+++ supplements/static/pls/js/common.js
@@ -208,7 +208,7 @@ $(document).ready(function(){
             $(".refund-amount").each(function (i, item) {
                 if ($(item).val()) {
                     data[($(item).data("line-id"))] = $(item).val();
-                    amount += parseInt($(item).val());
+                    amount += parseFloat($(item).val());
                 }
             });
         }
